Moved BrowserRouter to separate routes directory
