+++ src/App.tsx
@@ -1,0 +1,5 @@
+import { Routes } from './routes';
+
+export const App = () => {
+  return <Routes />;
+};
